requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.6.0
    # via django
django==4.1.5
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-filter
    #   django-guardian
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cors-headers==3.13.0
    # via -r requirements.in
django-filter==22.1
    # via -r requirements.in
django-guardian==2.4.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.2.2
    # via -r requirements.in
markdown==3.4.1
    # via -r requirements.in
pygments==2.14.0
    # via -r requirements.in
pyjwt==2.6.0
    # via djangorestframework-simplejwt
python-decouple==3.7
    # via -r requirements.in
pytz==2022.7.1
    # via djangorestframework
sqlparse==0.4.3
    # via django
